util: move fd_circq into util
