+++ src/util/circq/test_circq.c
@@ -1,4 +1,4 @@
-#include "../../util/fd_util.h"
+#include "../fd_util.h"
 #include "fd_circq.h"
 
 static void
